A little prettier looking! And better working too!
Obviously nothing is optimized or really pretty yet, but it's functional and can be used for what I needed. The only huge concern I have about the current functionality is that a final report isn't being generated.

I'm not sure how I'd tackle that one yet, but essentially I'd like to have the process send an email with the list of redirecting URLs. Maybe not an email though, that's just one option. Essentially I want each individual link result to be bundled together based on the group they were triggered by.